API/requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.0
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.3.9
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.10.1
python-dotenv==0.20.0
SQLAlchemy==1.4.35
typing_extensions==4.12.2
Werkzeug==2.1.1
zipp==3.8.0
